Migrated all remaining commands/queries to use validationcontext where applicable.

+++ src/PlexRequests/ApiRequests/Requests/Commands/DeleteMovieRequestCommandHandler.cs
@@ -1,8 +1,6 @@
-using System.Net;
 using System.Threading;
 using System.Threading.Tasks;
 using MediatR;
-using PlexRequests.Core.Exceptions;
 using PlexRequests.Core.Helpers;
 using PlexRequests.Core.Services;
 using PlexRequests.DataAccess;
@@ -12,7 +10,7 @@ using PlexRequests.DataAccess.Dtos;
 
 namespace PlexRequests.ApiRequests.Requests.Commands
 {
-    public class DeleteMovieRequestCommandHandler : AsyncRequestHandler<DeleteMovieRequestCommand>
+    public class DeleteMovieRequestCommandHandler : IRequestHandler<DeleteMovieRequestCommand, ValidationContext>
     {
         private readonly IMovieRequestService _requestService;
         private readonly IUnitOfWork _unitOfWork;
@@ -29,34 +27,41 @@ namespace PlexRequests.ApiRequests.Requests.Commands
             _claimsUserAccessor = claimsUserAccessor;
         }
         
-        protected override async Task Handle(DeleteMovieRequestCommand command, CancellationToken cancellationToken)
+        public async Task<ValidationContext> Handle(DeleteMovieRequestCommand command, CancellationToken cancellationToken)
         {
-            var request = await ValidateRequestExists(command);
+            var resultContext = new ValidationContext();
 
-            ValidateUserCanDeleteRequest(request);
+            var request = await ValidateRequestExists(command, resultContext);
+
+            ValidateUserCanDeleteRequest(request, resultContext);
+
+            if (!resultContext.IsSuccessful)
+            {
+                return resultContext;
+            }
 
             await _requestService.DeleteRequest(command.Id);
 
             await _unitOfWork.CommitAsync();
+
+            return resultContext;
         }
 
-        private void ValidateUserCanDeleteRequest(MovieRequestRow request)
+        private void ValidateUserCanDeleteRequest(MovieRequestRow request, ValidationContext resultContext)
         {
-            if (!request.UserId.Equals(_claimsUserAccessor.UserId))
+            if (request != null && !request.UserId.Equals(_claimsUserAccessor.UserId))
             {
-                throw new PlexRequestException("Unable to delete request", "Forbidden access to protected resource.",
-                    HttpStatusCode.Forbidden);
+                resultContext.AddError("Unable to delete request", "Forbidden access to protected resource.");
             }
         }
 
-        private async Task<MovieRequestRow> ValidateRequestExists(DeleteMovieRequestCommand command)
+        private async Task<MovieRequestRow> ValidateRequestExists(DeleteMovieRequestCommand command, ValidationContext resultContext)
         {
             var request = await _requestService.GetRequestById(command.Id);
 
             if (request == null)
             {
-                throw new PlexRequestException("Invalid request id", "A request for the given id was not found.",
-                    HttpStatusCode.NotFound);
+                resultContext.AddError("Invalid request id", "A request for the given id was not found.");
             }
 
             return request;
